+++ dune/copasi/model/diffusion_reaction.cc
@@ -7,6 +7,7 @@
  * file but a header which has to be included when compiling.
  */
 
+#include <dune/copasi/common/bit_flags.hh>
 #include <dune/copasi/common/data_context.hh>
 #include <dune/copasi/common/muparser_data_handler.hh>
 #include <dune/copasi/common/pdelab_expression_adapter.hh>
@@ -28,7 +29,7 @@ ModelDiffusionReaction<Traits>::ModelDiffusionReaction(
   std::shared_ptr<Grid> grid,
   const Dune::ParameterTree& config,
   GV grid_view,
-  ModelSetupPolicy setup_policy)
+  BitFlags<ModelSetup::Stages> setup_policy)
   : ModelBase(config)
   , _solver_logger(Logging::Logging::componentLogger(config, "solver"))
   , _components(config.sub("reaction").getValueKeys().size())
@@ -37,24 +38,7 @@ ModelDiffusionReaction<Traits>::ModelDiffusionReaction(
   , _grid(grid)
 {
   setup(setup_policy);
-
-  if constexpr (not Traits::is_sub_model) // todo fix this conditional for
-                                          // something else
-    if (setup_policy == ModelSetupPolicy::All) {
-      MuParserDataHandler<TIFFGrayscale<unsigned short>> mu_data_handler;
-      if (_config.hasSub("data"))
-        mu_data_handler.add_tiff_functions(_config.sub("data"));
-
-      auto initial_muparser = get_muparser_initial(_config, _grid_view, false);
-
-      for (auto&& mu_grid_function : initial_muparser) {
-        mu_data_handler.set_functions(mu_grid_function->parser());
-        mu_grid_function->set_time(current_time());
-        mu_grid_function->compile_parser();
-      }
-      set_initial(initial_muparser);
-      write_states();
-    }
+  write_states();
 
   _logger.debug("ModelDiffusionReaction constructed"_fmt);
 }
@@ -251,6 +235,34 @@ ModelDiffusionReaction<Traits>::setup_coefficient_vectors()
 
 template<class Traits>
 void
+ModelDiffusionReaction<Traits>::setup_initial_condition()
+{
+  // if this is a sub model, instantiation of the following instantiation will fail
+  if constexpr (not Traits::is_sub_model) {
+    // get TIFF data if available
+    MuParserDataHandler<TIFFGrayscale<unsigned short>> mu_data_handler;
+    if (_config.hasSub("data"))
+      mu_data_handler.add_tiff_functions(_config.sub("data"));
+
+    // configure math parsers for initial conditions on each component
+    auto initial_muparser = get_muparser_initial(_config, _grid_view, false);
+
+    for (auto&& mu_grid_function : initial_muparser) {
+      // make TIFF expression available in the parser
+      mu_data_handler.set_functions(mu_grid_function->parser());
+      mu_grid_function->set_time(current_time());
+      mu_grid_function->compile_parser();
+    }
+    // evaluate compiled expressions as initial conditions
+    set_initial(initial_muparser);
+  } else {
+    DUNE_THROW(NotImplemented,
+               "Initial condition setup is only available for complete models");
+  }
+}
+
+template<class Traits>
+void
 ModelDiffusionReaction<Traits>::setup_constraints()
 {
   _logger.debug("setup constraints"_fmt);
@@ -368,7 +380,14 @@ template<class Traits>
 void
 ModelDiffusionReaction<Traits>::setup_vtk_writer()
 {
-  _logger.debug("setup vtk writer"_fmt);
+  // only configure if config writer section exist
+  if(not _config.hasSub("writer"))
+  {
+    _logger.debug("skipping of setup vtk writer"_fmt);
+    return;
+  } else {
+    _logger.debug("setup vtk writer"_fmt);
+  }
 
   _writer = std::make_shared<W>(_grid_view, Dune::VTK::conforming);
   auto config_writer = _config.sub("writer");
@@ -400,21 +419,23 @@ ModelDiffusionReaction<Traits>::suggest_timestep(double dt)
 
 template<class Traits>
 void
-ModelDiffusionReaction<Traits>::setup(ModelSetupPolicy setup_policy)
+ModelDiffusionReaction<Traits>::setup(BitFlags<ModelSetup::Stages> setup_policy)
 {
-  if (setup_policy >= ModelSetupPolicy::GridFunctionSpace)
+  if (setup_policy.test(ModelSetup::Stages::GridFunctionSpace))
     setup_grid_function_space();
-  if (setup_policy >= ModelSetupPolicy::CoefficientVector)
+  if (setup_policy.test(ModelSetup::Stages::CoefficientVector))
     setup_coefficient_vectors();
-  if (setup_policy >= ModelSetupPolicy::Constraints)
+  if (setup_policy.test(ModelSetup::Stages::InitialCondition))
+    setup_initial_condition();
+  if (setup_policy.test(ModelSetup::Stages::Constraints))
     setup_constraints();
-  if (setup_policy >= ModelSetupPolicy::LocalOperator)
+  if (setup_policy.test(ModelSetup::Stages::LocalOperator))
     setup_local_operators();
-  if (setup_policy >= ModelSetupPolicy::GridOperator)
+  if (setup_policy.test(ModelSetup::Stages::GridOperator))
     setup_grid_operators();
-  if (setup_policy >= ModelSetupPolicy::Solver)
+  if (setup_policy.test(ModelSetup::Stages::Solver))
     setup_solvers();
-  if (setup_policy >= ModelSetupPolicy::Writer)
+  if (setup_policy.test(ModelSetup::Stages::Writer))
     setup_vtk_writer();
 }
 
@@ -445,7 +466,6 @@ ModelDiffusionReaction<Traits>::step()
     Dune::Logging::Logging::restoreCout();
 
   current_time() += dt;
-
   write_states();
 }
 
@@ -518,6 +538,9 @@ void
 ModelDiffusionReaction<Traits>::write_states(
   const std::map<std::size_t, ConstState>& states) const
 {
+  if (!_sequential_writer) // skip write if not setup
+    return;
+
   for (auto& [op, state] : _states) {
     auto& x = state.coefficients;
     auto& gfs = state.grid_function_space;