+++ dune/copasi/model/multidomain_diffusion_reaction.cc
@@ -7,6 +7,8 @@
  * file but a header which has to be included when compiling.
  */
 
+#include <dune/copasi/common/bit_flags.hh>
+
 #include <dune/copasi/common/muparser_data_handler.hh>
 #include <dune/copasi/model/multidomain_diffusion_reaction.hh>
 
@@ -19,7 +21,7 @@ template<class Traits>
 ModelMultiDomainDiffusionReaction<Traits>::ModelMultiDomainDiffusionReaction(
   std::shared_ptr<Grid> grid,
   const Dune::ParameterTree& config,
-  ModelSetupPolicy setup_policy)
+  BitFlags<ModelSetup::Stages> setup_policy)
   : ModelBase(config)
   , _solver_logger(Logging::Logging::componentLogger(config, "solver"))
   , _config(config)
@@ -28,27 +30,7 @@ ModelMultiDomainDiffusionReaction<Traits>::ModelMultiDomainDiffusionReaction(
   , _domains(config.sub("compartments").getValueKeys().size())
 {
   setup(setup_policy);
-
-  if (setup_policy == ModelSetupPolicy::All) {
-
-    MuParserDataHandler<TIFFGrayscale<unsigned short>> mu_data_handler;
-    if (_config.hasSub("data"))
-      mu_data_handler.add_tiff_functions(_config.sub("data"));
-
-    auto initial_muparser =
-      get_muparser_initial(_config, _grid->leafGridView(), false);
-
-    for (auto&& sd_grid_function : initial_muparser) {
-      for (auto&& mu_grid_function : sd_grid_function) {
-        mu_data_handler.set_functions(mu_grid_function->parser());
-        mu_grid_function->compile_parser();
-        mu_grid_function->set_time(current_time());
-      }
-    }
-    set_initial(initial_muparser);
-
-    write_states();
-  }
+  write_states();
 
   _logger.debug("ModelMultiDomainDiffusionReaction constructed"_fmt);
 }
@@ -171,8 +153,11 @@ ModelMultiDomainDiffusionReaction<Traits>::setup_grid_function_spaces()
     SubDomainGridView sub_grid_view = _grid->subDomain(domain).leafGridView();
 
     _logger.trace("create a sub model for compartment {}"_fmt, domain);
-    auto sub_model = std::make_shared<SubModel>(
-      _grid, model_config, sub_grid_view, ModelSetupPolicy::GridFunctionSpace);
+    auto sub_model =
+      std::make_shared<SubModel>(_grid,
+                                 model_config,
+                                 sub_grid_view,
+                                 ModelSetup::setup_grid_function_space);
     // extract grid function space from the sub model
     auto states = sub_model->states();
     for (auto& [op, state] : states) {
@@ -210,6 +195,31 @@ ModelMultiDomainDiffusionReaction<Traits>::setup_coefficient_vectors()
     else
       x = std::make_shared<X>(*gfs);
   }
+}
+
+template<class Traits>
+void
+ModelMultiDomainDiffusionReaction<Traits>::setup_initial_condition()
+{
+  // get TIFF data if available
+  MuParserDataHandler<TIFFGrayscale<unsigned short>> mu_data_handler;
+  if (_config.hasSub("data"))
+    mu_data_handler.add_tiff_functions(_config.sub("data"));
+
+  // configure math parsers for initial conditions on each component
+  auto initial_muparser =
+    get_muparser_initial(_config, _grid->leafGridView(), false);
+
+  for (auto&& sd_grid_function : initial_muparser) {
+    for (auto&& mu_grid_function : sd_grid_function) {
+      // make TIFF expression available in the parser
+      mu_data_handler.set_functions(mu_grid_function->parser());
+      mu_grid_function->compile_parser();
+      mu_grid_function->set_time(current_time());
+    }
+  }
+  // evaluate compiled expressions as initial conditions
+  set_initial(initial_muparser);
 }
 
 template<class Traits>
@@ -349,9 +359,17 @@ ModelMultiDomainDiffusionReaction<Traits>::setup_vtk_writer()
   _sequential_writer.resize(_domains);
 
   for (std::size_t i = 0; i < _domains; ++i) {
-
     const std::string compartement = compartments[i];
     auto& model_config = _config.sub(compartement);
+
+    // only configure those domains with writer section
+    if(not model_config.hasSub("writer"))
+    {
+      _logger.detail("skipping setup of vtk writer for domain '{}'"_fmt, compartement);
+      continue;
+    } else {
+      _logger.detail("setup vtk writer for domain '{}'"_fmt, compartement);
+    }
 
     int sub_domain_id =
       _config.sub("compartments").template get<int>(compartement);
@@ -360,7 +378,6 @@ ModelMultiDomainDiffusionReaction<Traits>::setup_vtk_writer()
 
     std::shared_ptr<W> writer =
       std::make_shared<W>(sub_grid_view, Dune::VTK::conforming);
-
     auto config_writer = model_config.sub("writer");
     std::string file_name =
       config_writer.template get<std::string>("file_name");
@@ -392,23 +409,26 @@ ModelMultiDomainDiffusionReaction<Traits>::suggest_timestep(double dt)
 
 template<class Traits>
 void
-ModelMultiDomainDiffusionReaction<Traits>::setup(ModelSetupPolicy setup_policy)
+ModelMultiDomainDiffusionReaction<Traits>::setup(
+  BitFlags<ModelSetup::Stages> setup_policy)
 {
   _logger.trace("setup operator started"_fmt);
 
-  if (setup_policy >= ModelSetupPolicy::GridFunctionSpace)
+  if (setup_policy.test(ModelSetup::Stages::GridFunctionSpace))
     setup_grid_function_spaces();
-  if (setup_policy >= ModelSetupPolicy::CoefficientVector)
+  if (setup_policy.test(ModelSetup::Stages::CoefficientVector))
     setup_coefficient_vectors();
-  if (setup_policy >= ModelSetupPolicy::Constraints)
+  if (setup_policy.test(ModelSetup::Stages::InitialCondition))
+    setup_initial_condition();
+  if (setup_policy.test(ModelSetup::Stages::Constraints))
     setup_constraints();
-  if (setup_policy >= ModelSetupPolicy::LocalOperator)
+  if (setup_policy.test(ModelSetup::Stages::LocalOperator))
     setup_local_operators();
-  if (setup_policy >= ModelSetupPolicy::GridOperator)
+  if (setup_policy.test(ModelSetup::Stages::GridOperator))
     setup_grid_operators();
-  if (setup_policy >= ModelSetupPolicy::Solver)
+  if (setup_policy.test(ModelSetup::Stages::Solver))
     setup_solvers();
-  if (setup_policy >= ModelSetupPolicy::Writer)
+  if (setup_policy.test(ModelSetup::Stages::Writer))
     setup_vtk_writer();
 }
 
@@ -584,12 +604,17 @@ void
 ModelMultiDomainDiffusionReaction<Traits>::write_states(
   const std::map<std::size_t, ConstState>& states) const
 {
+  if (_sequential_writer.empty()) // case where no writer was configured
+    return;
+
   const auto& compartments = _config.sub("compartments").getValueKeys();
 
   auto all_data = get_data_handler(states);
   for (std::size_t i = 0; i < _domains; ++i) {
-    const std::string compartement = compartments[i];
+    if (not _sequential_writer[i]) // skip not setup writers
+      continue;
 
+    const std::string compartement = compartments[i];
     for (auto& [op, state] : _states) {
       const auto& data = all_data[i].at(op);
       PDELab::vtk::OutputCollector<SW, DataHandler> collector(